Rename gdrive to shared_drive to align with PRT
